Move non-datamodel tagged nodes into something separate
