+++ src/roman_datamodels/maker_utils/__init__.py
@@ -4,6 +4,7 @@ from ._basic_meta import *  # noqa: F403
 from ._common_meta import *  # noqa: F403
 from ._datamodels import *  # noqa: F403
 from ._ref_files import *  # noqa: F403
+from ._tagged_nodes import *  # noqa: F403
 
 # These makers have special names to reflect the nature of their use in the pipeline
 SPECIAL_MAKERS = {
